fix: Workflow panic. Potential deadlock detected
> Potential deadlock detected

This error happens because workflow thread did not yield control via a Temporal
API call for over a second.

This change moves all the logic (that we thought is safe to execute in the
workflow directly) to be executed as a Local Activity.

Resolves LP:2085160

(cherry picked from commit f292d41ef3acdce1fd420c7cff31dd0ccba7343b)